admin_panel/django_api/movies_admin/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-split-settings==1.2.0
elastic-transport==8.12.0
elasticsearch==8.12.1
flake8==7.0.0
gunicorn==21.2.0
idna==3.7
mccabe==0.7.0
packaging==23.2
psycopg2-binary==2.9.8
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
uWSGI==2.0.24
djangorestframework==3.15.2